src/main/java/ru/ruthenium74/voteforrestaurant/web/restaurant/RestaurantRestController.java
package ru.ruthenium74.voteforrestaurant.web.restaurant;

import org.springframework.http.MediaType;
import org.springframework.lang.Nullable;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;
import ru.ruthenium74.voteforrestaurant.model.Restaurant;
import ru.ruthenium74.voteforrestaurant.repository.CrudRestaurantRepository;

import java.util.List;

@RestController
@RequestMapping(value = RestaurantRestController.REST_URL, produces = MediaType.APPLICATION_JSON_VALUE)
public class RestaurantRestController {
    public static final String REST_URL = "/rest/restaurants";

    private final CrudRestaurantRepository restaurantRepository;

    public RestaurantRestController(CrudRestaurantRepository restaurantRepository) {
        this.restaurantRepository = restaurantRepository;
    }

    @GetMapping
    public List<Restaurant> getAll(@RequestParam @Nullable boolean withDishes) {
        if (withDishes) {
            return restaurantRepository.getAllWithDishes();
        }
        return restaurantRepository.findAll();
    }
}
